# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: J-ONE HOTEL & LODGE API
  version: 1.0.0
  description: REST API powering the J-ONE HOTEL & LODGE public website and staff
    hotel management dashboard. The frontend is an independently hosted HTML/CSS/Vanilla-JS
    application communicating with this API over JSON.
paths:
...
  /api/admin/bookings/{lookup}/:
    get:
      operationId: admin_bookings_retrieve
      parameters:
      - in: path
        name: lookup
        schema:
          type: string
        required: true
      tags:
      - Admin · Bookings
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminBookingDetail'
          description: ''
    patch:
      operationId: admin_bookings_partial_update
      parameters:
      - in: path
        name: lookup
        schema:
          type: string
        required: true
      tags:
      - Admin · Bookings
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedAdminBookingDetailRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedAdminBookingDetailRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedAdminBookingDetailRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminBookingDetail'
          description: ''
...
  /api/bookings/{lookup}/:
    get:
      operationId: bookings_retrieve
      description: Resolve a booking by pk or reference + enforce object ownership.
      summary: Booking detail by id or reference
      parameters:
      - in: path
        name: lookup
        schema:
          type: string
        required: true
      tags:
      - Bookings
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingDetail'
          description: ''
...
  /api/rooms/:
    get:
      operationId: rooms_list
      summary: List active room types
      tags:
      - Rooms
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoomTypeList'
          description: ''
  /api/rooms/{slug}/:
    get:
      operationId: rooms_retrieve
      summary: Room type detail (images, amenities, applicable offers)
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - Rooms
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomTypeDetail'
          description: ''
...
components:
  schemas:
...
    AdminBookingDetail:
      type: object
      description: Row shape for the staff bookings table — one request for the whole
        page.
      properties:
        id:
          type: integer
          readOnly: true
        booking_reference:
          type: string
          maxLength: 40
        guest_name:
          type: string
          readOnly: true
        guest_phone:
          type: string
          readOnly: true
        guest_email:
          type: string
          readOnly: true
        room_type_name:
          type: string
          readOnly: true
        room_numbers:
          type: array
          items: {}
          readOnly: true
        check_in:
          type: string
          format: date
        check_out:
          type: string
          format: date
        nights:
          type: integer
          readOnly: true
        number_of_rooms:
          type: integer
          maximum: 9223372036854775807
          minimum: 1
          format: int64
        number_of_guests:
          type: string
          readOnly: true
        total_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        amount_paid:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        amount_due:
          type: string
          readOnly: true
        currency:
          type: string
          maxLength: 3
        status:
          $ref: '#/components/schemas/Status937Enum'
        payment_status:
          $ref: '#/components/schemas/PaymentStatusEnum'
        source:
          $ref: '#/components/schemas/Source350Enum'
        created_at:
          type: string
          format: date-time
          readOnly: true
        guest:
          allOf:
          - $ref: '#/components/schemas/GuestRead'
          readOnly: true
        room_assignments:
          type: array
          items:
            $ref: '#/components/schemas/BookingRoom'
          readOnly: true
        adults:
          type: integer
          maximum: 9223372036854775807
          minimum: 1
          format: int64
        children:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        price_per_night:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        subtotal:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        discount_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        extra_guest_fee_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        tax_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        fee_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        required_payment:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        refund_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        offer_title:
          type: string
          readOnly: true
        special_requests:
          type: string
        internal_notes:
          type: string
        cancellation_reason:
          type: string
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: While PENDING, inventory is held until this time.
        checked_in_at:
          type: string
          format: date-time
          nullable: true
        checked_out_at:
          type: string
          format: date-time
          nullable: true
        cancelled_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - amount_due
      - booking_reference
      - check_in
      - check_out
      - created_at
      - guest
      - guest_email
      - guest_name
      - guest_phone
      - id
      - nights
      - number_of_guests
      - offer_title
      - price_per_night
      - required_payment
      - room_assignments
      - room_numbers
      - room_type_name
      - subtotal
      - total_amount
      - updated_at
...
    BlankEnum:
      enum:
      - ''
    BookingDetail:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        booking_reference:
          type: string
          maxLength: 40
        guest:
          allOf:
          - $ref: '#/components/schemas/GuestRead'
          readOnly: true
        room_type_name:
          type: string
          readOnly: true
        room_type_slug:
          type: string
          readOnly: true
        room_assignments:
          type: array
          items:
            $ref: '#/components/schemas/BookingRoom'
          readOnly: true
        check_in:
          type: string
          format: date
        check_out:
          type: string
          format: date
        nights:
          type: integer
          readOnly: true
        number_of_rooms:
          type: integer
          maximum: 9223372036854775807
          minimum: 1
          format: int64
        adults:
          type: integer
          maximum: 9223372036854775807
          minimum: 1
          format: int64
        children:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        number_of_guests:
          type: integer
          readOnly: true
        price_per_night:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        subtotal:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        discount_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        extra_guest_fee_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        tax_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        fee_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        total_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        required_payment:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        amount_paid:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        amount_due:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
          readOnly: true
        refund_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        currency:
          type: string
          maxLength: 3
        offer_title:
          type: string
          readOnly: true
        status:
          $ref: '#/components/schemas/Status937Enum'
        payment_status:
          $ref: '#/components/schemas/PaymentStatusEnum'
        source:
          $ref: '#/components/schemas/Source350Enum'
        special_requests:
          type: string
        cancellation_reason:
          type: string
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: While PENDING, inventory is held until this time.
        checked_in_at:
          type: string
          format: date-time
          nullable: true
        checked_out_at:
          type: string
          format: date-time
          nullable: true
        cancelled_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        can_pay:
          type: string
          readOnly: true
        can_cancel:
          type: string
          readOnly: true
      required:
      - amount_due
      - booking_reference
      - can_cancel
      - can_pay
      - check_in
      - check_out
      - created_at
      - guest
      - id
      - nights
      - number_of_guests
      - offer_title
      - price_per_night
      - required_payment
      - room_assignments
      - room_type_name
      - room_type_slug
      - subtotal
      - total_amount
...
    BookingRoom:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        room_number:
          type: string
          readOnly: true
        floor:
          type: integer
          readOnly: true
        check_in:
          type: string
          format: date
        check_out:
          type: string
          format: date
      required:
      - check_in
      - check_out
      - floor
      - id
      - room_number
...
    GuestRead:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        full_name:
          type: string
          readOnly: true
        email:
          type: string
          format: email
          maxLength: 254
        phone:
          type: string
          maxLength: 20
        address:
          type: string
          maxLength: 255
        city:
          type: string
          maxLength: 100
        state:
          type: string
          maxLength: 100
        country:
          type: string
          maxLength: 100
        identification_type:
          oneOf:
          - $ref: '#/components/schemas/IdentificationTypeEnum'
          - $ref: '#/components/schemas/BlankEnum'
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - email
      - first_name
      - full_name
      - id
      - last_name
      - phone
...
    IdentificationTypeEnum:
      enum:
      - NATIONAL_ID
      - NIN
      - DRIVERS_LICENSE
      - PASSPORT
      - VOTERS_CARD
      - OTHER
      type: string
      description: |-
        * `NATIONAL_ID` - National ID
        * `NIN` - NIN Slip
        * `DRIVERS_LICENSE` - Driver's License
        * `PASSPORT` - International Passport
        * `VOTERS_CARD` - Voter's Card
        * `OTHER` - Other
...
    PatchedAdminBookingDetailRequest:
      type: object
      description: Row shape for the staff bookings table — one request for the whole
        page.
      properties:
        booking_reference:
          type: string
          minLength: 1
          maxLength: 40
        check_in:
          type: string
          format: date
        check_out:
          type: string
          format: date
        number_of_rooms:
          type: integer
          maximum: 9223372036854775807
          minimum: 1
          format: int64
        total_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        amount_paid:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        currency:
          type: string
          minLength: 1
          maxLength: 3
        status:
          $ref: '#/components/schemas/Status937Enum'
        payment_status:
          $ref: '#/components/schemas/PaymentStatusEnum'
        source:
          $ref: '#/components/schemas/Source350Enum'
        adults:
          type: integer
          maximum: 9223372036854775807
          minimum: 1
          format: int64
        children:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        price_per_night:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        subtotal:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        discount_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        extra_guest_fee_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        tax_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        fee_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        required_payment:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        refund_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        special_requests:
          type: string
        internal_notes:
          type: string
        cancellation_reason:
          type: string
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: While PENDING, inventory is held until this time.
        checked_in_at:
          type: string
          format: date-time
          nullable: true
        checked_out_at:
          type: string
          format: date-time
          nullable: true
        cancelled_at:
          type: string
          format: date-time
          nullable: true
...
    PaymentStatusEnum:
      enum:
      - UNPAID
      - PARTIALLY_PAID
      - PAID
      - PARTIALLY_REFUNDED
      - REFUNDED
      - FAILED
      type: string
      description: |-
        * `UNPAID` - Unpaid
        * `PARTIALLY_PAID` - Partially paid
        * `PAID` - Paid
        * `PARTIALLY_REFUNDED` - Partially refunded
        * `REFUNDED` - Refunded
        * `FAILED` - Payment failed
...
    RoomTypeDetail:
      type: object
      description: Full representation for the room detail page.
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 120
        slug:
          oneOf:
          - type: string
            maxLength: 140
            pattern: ^[-a-zA-Z0-9_]+$
          - type: string
            maxLength: 0
        short_description:
          type: string
          maxLength: 255
        base_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
          description: Price per night in hotel currency (NGN).
        max_guests:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        bed_type:
          type: string
          maxLength: 80
        bed_count:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        room_size:
          type: string
          description: e.g. "25 m²"
          maxLength: 60
        view:
          type: string
          maxLength: 80
        is_featured:
          type: boolean
        primary_image_url:
          type: string
          readOnly: true
        amenities:
          type: string
          readOnly: true
        description:
          type: string
        smoking_policy:
          $ref: '#/components/schemas/SmokingPolicyEnum'
        children_allowed:
          type: boolean
        extra_guest_allowed:
          type: boolean
        extra_guest_fee:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          description: Per extra guest, per night.
        images:
          type: string
          readOnly: true
      required:
      - amenities
      - base_price
      - id
      - images
      - name
      - primary_image_url
...
    RoomTypeList:
      type: object
      description: Lightweight representation for catalog/availability lists.
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 120
        slug:
          oneOf:
          - type: string
            maxLength: 140
            pattern: ^[-a-zA-Z0-9_]+$
          - type: string
            maxLength: 0
        short_description:
          type: string
          maxLength: 255
        base_price:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
          description: Price per night in hotel currency (NGN).
        max_guests:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        bed_type:
          type: string
          maxLength: 80
        bed_count:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        room_size:
          type: string
          description: e.g. "25 m²"
          maxLength: 60
        view:
          type: string
          maxLength: 80
        is_featured:
          type: boolean
        primary_image_url:
          type: string
          readOnly: true
        amenities:
          type: string
          readOnly: true
      required:
      - amenities
      - base_price
      - id
      - name
      - primary_image_url
    SmokingPolicyEnum:
      enum:
      - NON_SMOKING
      - SMOKING_ALLOWED
      type: string
      description: |-
        * `NON_SMOKING` - Non-smoking
        * `SMOKING_ALLOWED` - Smoking allowed
    Source350Enum:
      enum:
      - WEBSITE
      - WALK_IN
      - PHONE
      type: string
      description: |-
        * `WEBSITE` - Website
        * `WALK_IN` - Walk-in
        * `PHONE` - Phone
    Status937Enum:
      enum:
      - PENDING
      - CONFIRMED
      - CHECKED_IN
      - CHECKED_OUT
      - CANCELLED
      - EXPIRED
      - NO_SHOW
      type: string
      description: |-
        * `PENDING` - Pending payment
        * `CONFIRMED` - Confirmed
        * `CHECKED_IN` - Checked in
        * `CHECKED_OUT` - Checked out
        * `CANCELLED` - Cancelled
        * `EXPIRED` - Expired (unpaid)
        * `NO_SHOW` - No show
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT